pyproject.toml
[tool.poetry]
name = "cv-generator"
version = "0.1.0"
description = ""
authors = ["guimarca <guimarca@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
PyYAML = "^6.0.1"
Jinja2 = "^2.11.3"
WeasyPrint = "^57.0"
MarkupSafe = "2.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^5.2"
ruff = "^0.4.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
